Organize Wayland code
