+++ common.mk
@@ -151,6 +151,7 @@ SOURCES_FILTERED = $(SOURCES)
 endif
 
 OBJECTS = $(patsubst %.rc,%.o,$(patsubst %.c,%.o,$(patsubst %.mm,%.o,$(patsubst %.cc,%.o,$(SOURCES_FILTERED)))))
+D_FILES = $(patsubst %.o,%.d,$(OBJECTS))
 CFLAGS = $(STD_CFLAGS) $(CONFIG_CFLAGS) $(PLATFORM_CFLAGS) $(QT_CFLAGS) $(PKG_CFLAGS)
 CXXFLAGS = $(STD_CXXFLAGS) $(CONFIG_CXXFLAGS) $(PLATFORM_CXXFLAGS) $(QT_CFLAGS) $(PKG_CFLAGS)
 LDFLAGS = $(CONFIG_LDFLAGS) $(PLATFORM_LDFLAGS) $(QT_LDFLAGS) $(PKG_LDFLAGS)
@@ -177,7 +178,7 @@ CMAKEFLAGS = $(CONFIG_CMAKEFLAGS)
 all::
 
 clean::
-	rm -f *.d
+	rm -f $(D_FILES)
 
 FORCE: ;
